+++ test/tornado_tests/test_motor_database.py
@@ -15,7 +15,7 @@
 """Test Motor, an asynchronous driver for MongoDB and Tornado."""
 
 import unittest
-from test import env
+from test import SkipTest, env
 from test.tornado_tests import MotorTest
 
 import pymongo.database
@@ -167,6 +167,8 @@ class MotorDatabaseTest(MotorTest):
     @gen_test
     async def test_cursor_command(self):
         db = self.db
+        if not hasattr(db, "cursor_command"):
+            raise SkipTest("MOTOR-1169")
         db.test.drop()
 
         docs = [{"_id": i, "doc": i} for i in range(3)]
